+++ src/autofonce_lib/command_run.ml
@@ -48,7 +48,6 @@ let cmd =
   EZCMD.sub
     "run"
     (fun () ->
-       Printexc.record_backtrace true ;
        try
          let (p, tc, suite) = Testsuite.find () in
          let n = Testsuite.exec p tc suite in