+++ src/autofonce_lib/promote.mli
@@ -1,0 +1,16 @@
+(**************************************************************************)
+(*                                                                        *)
+(*  Copyright (c) 2023 OCamlPro SAS                                       *)
+(*                                                                        *)
+(*  All rights reserved.                                                  *)
+(*  This file is distributed under the terms of the GNU General Public    *)
+(*  License version 3.0, as described in the LICENSE.md file in the root  *)
+(*  directory of this source tree.                                        *)
+(*                                                                        *)
+(*                                                                        *)
+(**************************************************************************)
+
+val print_actions :
+  not_exit:bool ->
+  keep_old:bool ->
+  Buffer.t -> Types.action list -> unit